feat(mindmap): add deleteSelectedElements function , use when cut and delete
