+++ .changeset/sour-adults-remain.md
@@ -1,0 +1,5 @@
+---
+'@plait/mindmap': patch
+---
+
+add deleteSelectedElements function , use when cut and delete
